refactor(chat): extract turn controller + consolidate catalog status
Two fixes to unblock CI on development:

1. file-lines check (bun run check):
   ChatContainer.tsx was 518 lines, over the 500 budget. Pulled
   useChatTurnController + buildInitialConversationTitle (and their
   types) into frontend/features/chat/hooks/use-chat-turn-controller.ts.
   ChatContainer.tsx drops to 393 lines; new hook is 160 lines.
   Same behavior, same wiring — just relocated.

2. React Doctor warning (fail-on: warning):
   LandingState was taking 4 boolean-like props (isLoading,
   isCatalogLoading, isCatalogError, isComposerBlocked) which
   React Doctor flags as "stacked flags". Consolidated the two
   mutually-exclusive catalog flags into a single CatalogStatus
   discriminated union ('loading' | 'error' | 'ready'), per the
   repo's discriminated-unions-over-booleans rule. The wrapper
   components (ChatComposer, ModelSelectorPopover) continue to
   accept the older booleans, so they're derived inline at the
   call site rather than rippling through the vendored composer
   surface.

Verified:
- biome check: clean
- tsc --noEmit: clean (ignoring pre-existing fumadocs .source/server)
- vitest features/chat: 59/59 passing
- node scripts/check-file-lines.mjs: OK
- node scripts/check-nesting.mjs: OK
